feat: 장바구니 페이지와 주문 확인 페이지에서 고통으로 쓰이는 컴포넌트 공용화
