+++ src/components/CartItem/CountButtonContainer.tsx
@@ -1,16 +1,17 @@
 import { css } from '@emotion/react';
 import { useSetRecoilState } from 'recoil';
 
+import { updateItemQuantity } from '@/apis/cartItem';
+import { MINUS, PLUS } from '@/assets/images';
 import { THEME } from '@/constants/theme';
+import { cartItemsState } from '@/recoil/cartItems/atoms';
 import { CartItemProps } from '@/types/cartItem';
-import { updateItemQuantity } from '@apis/cartItem';
-import { MINUS, PLUS } from '@assets/images';
-import { cartItemsState } from '@recoil/cartItems/atoms';
 
-interface CartItemMainSectionProps {
+interface CountButtonContainerProps {
   item: CartItemProps;
 }
-const CartItemMainSection = ({ item }: CartItemMainSectionProps) => {
+
+const CountButtonContainer = ({ item }: CountButtonContainerProps) => {
   const setCartItems = useSetRecoilState(cartItemsState);
 
   const handleDecrementQuantity = async () => {
@@ -49,46 +50,24 @@ const CartItemMainSection = ({ item }: CartItemMainSectionProps) => {
   };
 
   return (
-    <div css={cartItemBody}>
-      <img css={image} src={item.product.imageUrl} width={112} height={112} />
-      <div css={cartItemInfoWrapper}>
-        <span>{item.product.name}</span>
-        <span css={price}>{item.product.price.toLocaleString('ko-KR')}원</span>
-        <div css={countWrapper}>
-          <button
-            id="minus-button"
-            css={countButton(item.quantity === 1)}
-            onClick={handleDecrementQuantity}
-            disabled={item.quantity === 1}
-          >
-            <img src={MINUS} alt={`${item.product.name}-minus`} />
-          </button>
-          <span>{item.quantity}</span>
-          <button css={countButton()} onClick={handleIncrementQuantity}>
-            <img src={PLUS} alt={`${item.product.name}-plus`} />
-          </button>
-        </div>
-      </div>
+    <div css={countWrapper}>
+      <button
+        id="minus-button"
+        css={countButton(item.quantity === 1)}
+        onClick={handleDecrementQuantity}
+        disabled={item.quantity === 1}
+      >
+        <img src={MINUS} alt={`${item.product.name}-minus`} />
+      </button>
+      <span>{item.quantity}</span>
+      <button css={countButton()} onClick={handleIncrementQuantity}>
+        <img src={PLUS} alt={`${item.product.name}-plus`} />
+      </button>
     </div>
   );
 };
 
-export default CartItemMainSection;
-
-const cartItemBody = css`
-  display: flex;
-  gap: 20px;
-`;
-
-const image = css`
-  border-radius: 8px;
-`;
-
-const cartItemInfoWrapper = css`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-`;
+export default CountButtonContainer;
 
 const countWrapper = css`
   width: 80px;
@@ -97,11 +76,6 @@ const countWrapper = css`
   justify-content: space-between;
   align-items: center;
   gap: 8px;
-`;
-
-const price = css`
-  font-weight: 700;
-  font-size: 24px;
 `;
 
 const countButton = (isDisabled?: boolean) => css`
