+++ src/index.css
@@ -114,8 +114,10 @@ body {
 }
 
 ol,
-ul {
+ul,
+li {
   list-style: none;
+
 }
 
 blockquote,
